Load databases once into class level variables.
Instead of loading the databases every time into every instance, which takes time,
they are loaded once and stored in the AutoTST_Reaction class. This may not be the
best place to store it, but it's closest to where things are now and minimizes the
refactoring of this single commit.

One side-effect is that all three reaction families are now loaded from the RMG-database,
so when generating reactions you have to restrict it using the only_families parameter.