Don't calculate TS geometry until required.
I'm using a @property decorator and a getter function to replace the `.ts`
attribute. The TS is now stored internally as `._ts` but is only calculated
if you ask for `.ts`. This should not break downstream code that wants to
access the TS via `.ts`, but will speed up code that doesn't need it.

If there are side-effects of calling create_ts_geometries() that are relied
upon, then that might break things, but it should be fixed by removing them and
making them explicit.