Don't generate distance_data until it's required.
Like the previous commit, this shouldn't break downstream code
that expects to get distance data from `.distance_data`
but will speed up code that doesn't need it.
It's stored internally as `._distance_data` but only
updated with an explicit call to `.generate_distance_data()`
or when `.distance_data` is accessed for the first time.