Fixed our treatment of HR for TSs
Before, we were optimizing to a saddle point after each optimization which was easily prone to errors. After looking through some of Pierre's old log files, we saw that he was optimizing to a minima at each set BUT was keeping the reaction center frozen along each step. This commit makes that change.