Reworked the way we find center bonds
Before, we were looking at if both ends of an RMG bond have labeled atoms, but that would result in many bonds incorrectly being labeled as in the reaction center in the case of intra_h_migration. This is because RMG was updated to label all atoms in the ring. Our old workflow would look for all labeled atoms (regardless of if they were the labels of interest). This commit fixes that.